Assortment of test juggling for clang.
llvm-svn: 125087

+++ clang-tests/gcc-4_2-testsuite/src/g++.dg/lookup/new1.C
@@ -4,10 +4,8 @@
 
 int main() {
   int i;
-  void* operator new(unsigned s, int* p);
+  void* operator new(unsigned long s, int* p);
   int* e = new(&i) int;                    // { dg-error "no matching function" }
   int* f = new int;
   return 0;
 }
-
-// { dg-excess-errors "operator new" }
